Allocate a number of buffers and use them to decompress tile data
Instead of allocating as many and as large buffers as necessary to
decompress tiles on demand, pre-allocate an amount (128MB) of memory and
use the fzstd stream interface to decompress into that.

Once these buffers are full, tile updates block on slurp. This markedly
improves memory use and behaviour, in Firefox at least.

Signed-off-by: Chris Lord <chrislord.net@gmail.com>
Change-Id: Ibacc3c330d25e0bd7c755933e9a9329d1d039033